Move safe_sequences_wrapper.h to find_kmers_specific
