+++ src/find_kmers_specific/safe_sequences_wrapper.h
